pyproject.toml
[tool.poetry]
name = "frontier-scans-cleanup"
version = "0.2.1"
description = "A script to fix metadata and reorganize film scans from Fujifilm Frontier scanners."
authors = ["Bobo Li <bobo@photolaundrysf.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
Wand = "^0.6.11"
PyExifTool = "^0.5.6"

[tool.poetry.scripts]
frontier-clean-c4c5 = "frontier_scans_cleanup.cleanup_c4c5:cli"
frontier-clean-ms01 = "frontier_scans_cleanup.cleanup_ms01:cli"

[tool.ruff]
line-length = 79

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
